backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-decouple==3.8
python-dotenv==1.0.0
pydantic==2.5.0
email-validator==2.1.0
httpx==0.28.1